+++ fuzz/fuzz_targets/fuzz_block_eval.rs
@@ -1,7 +1,9 @@
 #![no_main]
-use libfuzzer_sys::fuzz_target;
-extern crate all_is_cubes;
+#![allow(clippy::single_match)]
 
+use libfuzzer_sys::fuzz_target;
+
+extern crate all_is_cubes;
 use all_is_cubes::block::Block;
 
 fuzz_target!(|block: Block| {
@@ -10,22 +12,12 @@ fuzz_target!(|block: Block| {
 
     let sink = all_is_cubes::listen::Sink::new();
 
-    match (block.evaluate(), block.listen(sink.listener())) {
-        (Ok(evaluated), Ok(())) => {
+    match block.evaluate_and_listen(sink.listener()) {
+        Ok(evaluated) => {
             evaluated.consistency_check();
         }
-        (Err(_), Err(_) | Ok(())) => {
+        Err(_) => {
             // Errors are an expected possibility; this fuzz test is looking for no panic.
-            //
-            // It should never be the case that listen() fails when evaluate() succeeds.
-            // (The case when evaluate() only fails is when a `BlockDef` has a broken ref
-            // inside it.)
-
-            // TODO: something like assert_eq!(eval_err, listen_err)
-            // (but for now there are differences in _which_ URef is checked first)
-        }
-        (eval_result, listen_result) => {
-            panic!("Bad outcome:\neval {eval_result:?}\nlisten {listen_result:?}");
         }
     }